+++ src/ui/dashboard/menu.tsx
@@ -1,9 +1,15 @@
 import Link from "next/link"
 
-const links = [
-    { href: "/dashboard/samples", label: "List" },
-    { href: "/dashboard/samples/add", label: "Add" },
-]
+const links = {
+    samples: [
+        { href: "/dashboard/samples", label: "List" },
+        { href: "/dashboard/samples/add", label: "Add" },
+    ],
+    zones: [
+        { href: "/dashboard/zones", label: "List" },
+        { href: "/dashboard/zones/add", label: "Add" },
+    ],
+}
 
 export const Menu = () => {
     return (
@@ -11,14 +17,16 @@ export const Menu = () => {
             <h2 className="text-lg text-neutral-800 font-medium">
                 <Link href="/dashboard">Dashboard</Link>
             </h2>
-            <ul className="text-neutral-500 space-y-1">
-                <li className="text-neutral-600 font-medium">Samples</li>
-                {links.map(({ href, label }, key) => (
-                    <li className="ml-3" key={key}>
-                        <Link href={href}>{label}</Link>
-                    </li>
-                ))}
-            </ul>
+            {Object.entries(links).map(([key, links]) => (
+                <ul className="text-neutral-500 space-y-1" key={key}>
+                    <li className="text-neutral-600 font-medium capitalize">{key}</li>
+                    {links.map(({ href, label }, key) => (
+                        <li className="ml-3" key={key}>
+                            <Link href={href}>{label}</Link>
+                        </li>
+                    ))}
+                </ul>
+            ))}
         </aside>
     )
 }
