+++ src/ui/dashboard/zones/add-zones.tsx
@@ -1,16 +1,16 @@
 "use client"
-import { AddZonesActionState } from "@/lib/@types/types"
-import { addZonesAction } from "@/lib/actions"
-import { getPlantsByUser } from "@/lib/services/dashboard"
-import { Plants } from "@prisma/client"
-import { useSession } from "next-auth/react"
-import { useEffect, useState } from "react"
+import Image from "next/image"
 import { useFormState } from "react-dom"
+import { useEffect, useState } from "react"
+import { useSession } from "next-auth/react"
+import { Plants } from "@prisma/client"
+import { addZonesAction } from "@/lib/actions"
+import { getPlantsByUser } from "@/lib/services/plants"
+import { AddZonesActionState } from "@/lib/@types/types"
 import { Form } from "@halvaradop/ui-form"
 import { Input } from "@halvaradop/ui-input"
 import { Label } from "@halvaradop/ui-label"
 import { Button } from "@halvaradop/ui-button"
-import Image from "next/image"
 import arrowDown from "@/public/arrow.svg"
 
 export const AddZone = () => {
